All tests work!
