--- tests/index.test.ts.dontTest
@@ -1,9 +1,0 @@
-
-
-// test("index route works", done => {
-//   request(app)
-//     .get("/")
-//     .expect("Content-Type", /json/)
-//     .expect({ name: "frodo" })
-//     .expect(200, done);
-// });
